Parameterized Join Topology and added test for LEFT and OUTER.
Bare with me:
I know this is not what we normally want to teach people but I found this example very instructive, for several reasons.
Lambdas hide a lot of what is going on behind syntactic sugar, but sometimes that shit surfaces again and makes life difficult.
It is therefore useful to show people what actually happens under the covers and how one can replace the lambdas with the
actual methods generated. In fact, the method signature itself of the join methods mentions the underlying classes.

Combining all three join methods in one topology has the advantage to test and show the differences very quickly without further code changes,
and that has didactic value as well.

But I needed to get rid of all the lambdas because Java cannot deduce anymore from the methods what the generic types are, and so complains.
By using an explicit ValueJoiner and ForeachAction I can circumvent that.